fix: simplify vesting exported interfaces
The use of the "action" interfaces is to allow account wrappers
without creating cyclic package imports. But the mechanism only
allows a single sequence of tests, not an arbitrary graph.
Therefore we simplify (for now) that a ClawbackVestingAccount
is the only account type that supports return-grants. If we add,
say, PeriodicVestingAccount in the future, we'll add the ReturnGrants
method to the GrantAccount interface, etc.